refactor: S3 접근 계층을 global/infra/storage로 이동하고 조회,삭제 기능 추가
