+++ Pokade/src/main/java/com/pokade/domain/user/controller/ProfileImageController.java
@@ -1,0 +1,4 @@
+package com.pokade.domain.user.controller;
+
+public class ProfileImageController {
+}
