+++ Pokade/src/main/java/com/pokade/domain/user/service/ProfileImageService.java
@@ -1,0 +1,4 @@
+package com.pokade.domain.user.service;
+
+public class ProfileImageService {
+}
